main: Add audio record fragment
The audio record fragment is used to record an audio stream. By doing this
we can get the current amplitude which is used to mesure the ambient noise
in dB. This action is limited to a maximum of 30 seconds, meaning that
after this time period the audio stream will close. After each recording
the average noise level mesured is presented on the screen.

Reference-issue: TG2020S-7
Signed-off-by: João Lourenço <jlourenco5691@gmail.com>